requirements.txt
fastapi
langchain-groq
langchain_community
langgraph
uvicorn[standard]
cloudinary
scikit-learn
numpy
pydantic
python-dotenv
requests
python-multipart
pypdf
sentence-transformers>=2.2.0
langchain-core>=0.1.0
unstructured
pathlib